requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audio-processing==0.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
crcmod==1.7
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.1
flask-sock==0.7.0
google==3.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.86.0
protobuf==6.31.1
pvporcupine==3.0.5
py==1.11.0
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.9.9
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
pytz==2020.5
requests==2.32.4
retry==0.9.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
srt==3.5.3
tos==2.8.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
volcengine==1.0.188
vosk==0.3.45
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
